move insightsProvider to core
