+++ thirdeye-server/src/main/java/ai/startree/thirdeye/resources/AlertResource.java
@@ -14,16 +14,18 @@
 package ai.startree.thirdeye.resources;
 
 import static ai.startree.thirdeye.spi.ThirdEyeStatus.ERR_CRON_INVALID;
+import static ai.startree.thirdeye.spi.ThirdEyeStatus.ERR_UNKNOWN;
 import static ai.startree.thirdeye.spi.util.SpiUtils.optional;
 import static ai.startree.thirdeye.util.ResourceUtils.ensure;
 import static ai.startree.thirdeye.util.ResourceUtils.ensureExists;
 import static ai.startree.thirdeye.util.ResourceUtils.respondOk;
+import static ai.startree.thirdeye.util.ResourceUtils.serverError;
 
 import ai.startree.thirdeye.auth.ThirdEyePrincipal;
 import ai.startree.thirdeye.core.AlertCreater;
 import ai.startree.thirdeye.core.AlertDeleter;
 import ai.startree.thirdeye.core.AlertEvaluator;
-import ai.startree.thirdeye.core.AlertInsightsProvider;
+import ai.startree.thirdeye.alert.AlertInsightsProvider;
 import ai.startree.thirdeye.mapper.AlertApiBeanMapper;
 import ai.startree.thirdeye.mapper.ApiBeanMapper;
 import ai.startree.thirdeye.spi.api.AlertApi;
@@ -53,6 +55,7 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -151,9 +154,15 @@ public class AlertResource extends CrudResource<AlertApi, AlertDTO> {
       @PathParam("id") final Long id) {
     final AlertDTO dto = get(id);
 
-    final AlertInsightsApi insights = alertInsightsProvider.getInsights(dto);
-
-    return Response.ok(insights).build();
+    try {
+      final AlertInsightsApi insights = alertInsightsProvider.getInsights(dto);
+      return Response.ok(insights).build();
+    } catch (final WebApplicationException e) {
+      throw e;
+    } catch (final Exception e) {
+      // can do better exception handling - see handleAlertEvaluationException
+      throw serverError(ERR_UNKNOWN, e);
+    }
   }
 
   @Path("{id}/run")
